move debug stat spec to playwright
