+++ playwright/debug-stat-loading.spec.js
@@ -1,0 +1,33 @@
+import { test } from "./playwright/fixtures/commonSetup.js";
+
+test("debug stat loading issue", async ({ page }) => {
+  const logs = [];
+
+  // Capture all console messages
+  page.on("console", (msg) => {
+    logs.push(`[${msg.type()}] ${msg.text()}`);
+  });
+
+  // Go to the page
+  await page.goto("/src/pages/battleCLI.html");
+
+  // Try to start a match
+  const startBtn = page.locator("#start-match-button");
+  if ((await startBtn.count()) > 0) {
+    await startBtn.click();
+  }
+
+  // Wait a moment for logs
+  await page.waitForTimeout(3000);
+
+  // Print all console messages
+  console.log("=== Console Messages ===");
+  logs.forEach((log) => console.log(log));
+
+  // Check if stat list is populated
+  const statElements = await page.locator("#cli-stats .cli-stat").count();
+  console.log(`Found ${statElements} stat elements`);
+
+  const statsContent = await page.locator("#cli-stats").innerHTML();
+  console.log("Stats container content:", statsContent);
+});
